Add card, quote, call to action, and footer to the page
Complete the landing page adding all desired content. As a reflection
I would like to try this project again with a different approach, and
more knowledge on how to make more accessable on smaller screens. I
also learned the importance of laying out the html first. I didn't
do this and a lot of css could have been combinded which could have
saved a lot of time.